bound to uint64

+++ ethereum/contracts/test/Channels.t.sol
@@ -1862,9 +1862,9 @@ contract HoprChannelsTest is Test, ERC1820RegistryFixtureTest, CryptoUtils, Hopr
         vm.clearMockedCalls();
     }
 
-    function testFuzz_DomainSeparator(uint256 newChainId) public {
+    function testFuzz_DomainSeparator(uint64 newChainId) public {
         // chain ID must be less than 2^64 - 1
-        newChainId = bound(newChainId, 1, 0xFFFFFFFFFFFFFFFF);
+        newChainId = uint64(bound(uint256(newChainId), 1, 0xFFFFFFFFFFFFFFFF));
         vm.assume(newChainId != block.chainid);
         bytes32 domainSeparatorOnDeployment = hoprChannels.domainSeparator();
 
@@ -1890,9 +1890,9 @@ contract HoprChannelsTest is Test, ERC1820RegistryFixtureTest, CryptoUtils, Hopr
         assertTrue(hoprChannels.domainSeparator() != domainSeparatorOnDeployment);
     }
 
-    function testFuzz_LedgerDomainSeparator(uint256 newChainId) public {
+    function testFuzz_LedgerDomainSeparator(uint64 newChainId) public {
         // chain ID must be less than 2^64 - 1
-        newChainId = bound(newChainId, 1, 0xFFFFFFFFFFFFFFFF);
+        newChainId = uint64(bound(uint256(newChainId), 1, 0xFFFFFFFFFFFFFFFF));
         vm.assume(newChainId != block.chainid);
         bytes32 ledgerDomainSeparatorOnDeployment = hoprChannels.ledgerDomainSeparator();
